新品fragment的数据加载

+++ src/cn/ucai/fulicenter/bean/NewGoodBean.java
@@ -50,7 +50,7 @@ public class NewGoodBean implements Serializable{
     private String colorName;
     private String colorCode;
     private String colorUrl;
-    private int addTime;
+    private long addTime;
 
     public int getId() {
         return id;
@@ -188,7 +188,7 @@ public class NewGoodBean implements Serializable{
         this.colorUrl = colorUrl;
     }
 
-    public int getAddTime() {
+    public long getAddTime() {
         return addTime;
     }
